Added CDSAdaptorTest with updated osgi verison - updated osgi from 2012 to 2020 version - enabled for all jdk versions - modified Testbundle and Weavinghook from jar to source and manifest - removed outdated cds jars - modified com.ibm.* to org.openj9.test.* - Related to issue: runtimes/backlog 394
[ci skip]
