+++ test/functional/cmdLineTests/CDSAdaptorTest/readme.txt
@@ -1,3 +1,25 @@
+*******************************************************************************
+*  Copyright (c) 2001, 2020 IBM Corp. and others
+*
+*  This program and the accompanying materials are made available under
+*  the terms of the Eclipse Public License 2.0 which accompanies this
+*  distribution and is available at https://www.eclipse.org/legal/epl-2.0/
+*  or the Apache License, Version 2.0 which accompanies this distribution and
+*  is available at https://www.apache.org/licenses/LICENSE-2.0.
+*
+*  This Source Code may also be made available under the following
+*  Secondary Licenses when the conditions for such availability set
+*  forth in the Eclipse Public License, v. 2.0 are satisfied: GNU
+*  General Public License, version 2 with the GNU Classpath
+*  Exception [1] and GNU General Public License, version 2 with the
+*  OpenJDK Assembly Exception [2].
+*
+*  [1] https://www.gnu.org/software/classpath/license.html
+*  [2] http://openjdk.java.net/legal/assembly-exception.html
+*
+*  SPDX-License-Identifier: EPL-2.0 OR Apache-2.0 OR GPL-2.0 WITH Classpath-exception-2.0 OR LicenseRef-GPL-2.0 WITH Assembly-exception
+*******************************************************************************
+
 This test suite is the place holder for adding CDS Adaptor tests.
 Command line tests present in cdsadaptortest.xml use combination of java classes and OSGi bundles to achieve their purpose.
 Main purpose of java class is to start OSGi framework and install required bundles. 
@@ -5,7 +27,7 @@ But they may not be limited to this task only.
 
 Few guidelines/assumptions when writing/updating tests placed in this project:
 
-1) Java class are placed in com.ibm.cdsadaptortest package.
+1) Java class are placed in org.openj9.test.cdsadaptortest package.
 
 2) OSGi bundles required by the java class are created outside this project and
    only the bundle jar (including the source code) is committed to this project.
@@ -23,13 +45,13 @@ Few guidelines/assumptions when writing/updating tests placed in this project:
 	|
 	|--FrameworkBundles: this directory contains bundles required by the framework. eg CDS Adaptor bundle.
 	|
-	|--org.eclipse.osgi_<version>.jar: this is Equinox jar file, an OSGi framework implementation.
+	|--org.eclipse.osgi_<version>.jar: this is Equinox jar file, an OSGi framework implementation, and it contains cds adaptor.
 	|
-	|--Test1Bundles: this directory contains bundles required by java class Test1.
+	|--TestBundle: this directory contains bundle source code required by java class TestBundle.
 	|
-	|--Test2Bundles: this directory contains bundles required by java class Test2.
+	|--WeavinghookTest: this directory contains bundle source code required by java class WeavinghookTest.
 	
-5) If a new version of Equinox jar or CDS Adaptor is checked in, their version number in cdsadaptortest.xml needs to be updated as well.
+5) If a new version of Equinox jar is checked in, their version number in cdsadaptortest.xml needs to be updated as well.
    Failing to do this will cause tests to fail.
    Equinox builds are available at: http://download.eclipse.org/equinox/
