+++ test/functional/cmdLineTests/CDSAdaptorTest/resources/TestBundle/MANIFEST.MF
@@ -17,8 +17,8 @@ Manifest-Version: 1.0
 Bundle-Vendor: IBM
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-Name: Testplugin
-Bundle-SymbolicName: com.ibm.testbundle
-Bundle-Activator: com.ibm.testbundle.Activator
+Bundle-SymbolicName: org.openj9.test.testbundle
+Bundle-Activator: org.openj9.test.testbundle.Activator
 Bundle-Version: 1.0.0._v1
 Import-Package: org.osgi.framework;version="1.3.0"
 Bundle-ManifestVersion: 2
