--- test/functional/cmdLineTests/CDSAdaptorTest/resources/TestBundle/src/com/ibm/testbundle/SomeMessageV2.java
@@ -1,8 +1,0 @@
-package com.ibm.testbundle;
-
-public class SomeMessageV2 {
-
-	public void printMessage() {
-		System.out.println("A message from modified class");
-	}
-}
