+++ test/functional/cmdLineTests/CDSAdaptorTest/resources/WeavinghookTest/MANIFEST.MF
@@ -17,8 +17,8 @@ Manifest-Version: 1.0
 Bundle-Vendor: IBM
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-Name: Weavinghooktest
-Bundle-SymbolicName: com.ibm.weavinghooktest
-Bundle-Activator: com.ibm.weavinghooktest.Activator
+Bundle-SymbolicName: org.openj9.test.weavinghooktest
+Bundle-Activator: org.openj9.test.weavinghooktest.Activator
 Bundle-Version: 1.0.0._v1
 Import-Package: org.osgi.framework;version="1.3.0",org.osgi.framework.
  hooks.weaving;version="1.0.0"
