+++ test/functional/cmdLineTests/CDSAdaptorTest/resources/readme.txt
@@ -20,7 +20,7 @@
 *  SPDX-License-Identifier: EPL-2.0 OR Apache-2.0 OR GPL-2.0 WITH Classpath-exception-2.0 OR LicenseRef-GPL-2.0 WITH Assembly-exception
 *******************************************************************************
 
-This file explains the working of com.ibm.cdsadaptortest.CDSAdaptorOrphanTest.java
+This file explains the working of org.openj9.test.cdsadaptortest.CDSAdaptorOrphanTest.java
 This test verifies that a class modified by a weaving hook is stored as ORPHAN in shared class cache.
 
 Usage: 
@@ -35,38 +35,38 @@ Only describes the arguments accepted by this program. For complete command, ple
 Operation:
 It requires two bundles for its operation:
 
-1) com.ibm.testbundle: This is a simple bundle that consists of two classes:
-	com.ibm.testbundle.SomeMessageV1: Loaded by the bundle when it starts up.
+1) org.openj9.test.testbundle: This is a simple bundle that consists of two classes:
+	org.openj9.test.testbundle.SomeMessageV1: Loaded by the bundle when it starts up.
 									  It has a method printMessage() that prints a message M1.
 									  This method is called by the bundle when it starts up.
-	com.ibm.testbundle.SomeMessageV2: Identical to com.ibm.testbundle.SomeMessageV2 except that its printMessage() method prints a message M2.
-									  Its class bytes are used by weaving hook to replace class bytes of com.ibm.testbundle.SomeMessageV1.
+	org.openj9.test.testbundle.SomeMessageV2: Identical to org.openj9.test.testbundle.SomeMessageV2 except that its printMessage() method prints a message M2.
+									  Its class bytes are used by weaving hook to replace class bytes of org.openj9.test.testbundle.SomeMessageV1.
 	
-2) com.ibm.weavinghooktest: This is the weaving hook that transforms class bytes of com.ibm.testbundle.SomeMessageV1 to com.ibm.testbundle.SomeMessageV2.
+2) org.openj9.test.weavinghooktest: This is the weaving hook that transforms class bytes of org.openj9.test.testbundle.SomeMessageV1 to org.openj9.test.testbundle.SomeMessageV2.
 
-As mentioned in "Usage" section, com.ibm.cdsadaptortest.CDSAdaptorOrphanTest accepts an argument "-ignoreWeavingHookBundle". 
-If this argument is specified com.ibm.weavinghooktest bundle is not installed. By default, both the bundles mentioned above are installed.
+As mentioned in "Usage" section, org.openj9.test.cdsadaptortest.CDSAdaptorOrphanTest accepts an argument "-ignoreWeavingHookBundle". 
+If this argument is specified org.openj9.test.weavinghooktest bundle is not installed. By default, both the bundles mentioned above are installed.
 
-If "-ignoreWeavingHookBundle" is not specified, weaving hook replaces class bytes of com.ibm.testbundle.SomeMessageV1 with com.ibm.testbundle.SomeMessageV2.
-Hence the output message is M2 printed by printMessage() com.ibm.testbundle.SomeMessageV2 class.
+If "-ignoreWeavingHookBundle" is not specified, weaving hook replaces class bytes of org.openj9.test.testbundle.SomeMessageV1 with org.openj9.test.testbundle.SomeMessageV2.
+Hence the output message is M2 printed by printMessage() org.openj9.test.testbundle.SomeMessageV2 class.
 
 If "-ignoreWeavingHookBundle" is specified, weaving hook is not installed.
-Hence the output message is M1 printed by printMessage() com.ibm.testbundle.SomeMessageV1 class.
+Hence the output message is M1 printed by printMessage() org.openj9.test.testbundle.SomeMessageV1 class.
 
 How to run:
 This test is available in the builds at jvmtest/VM/cdsadaptortest/cdsadaptortest.jar.
 Unzip the cdsadaptortest.jar and run following command:
 	
-	java -Xshareclasses -cp .:./org.eclipse.osgi_3.16.100.v20200904-1304.jar com.ibm.cdsadaptortest.CDSAdaptorOrphanTest -frameworkBundleLocation ./FrameworkBundles -testBundleLocation ./CDSAdaptorOrphanTestBundles
+	java -Xshareclasses -cp .:./org.eclipse.osgi_3.16.100.v20200904-1304.jar org.openj9.test.cdsadaptortest.CDSAdaptorOrphanTest -frameworkBundleLocation ./FrameworkBundles -testBundleLocation ./CDSAdaptorOrphanTestBundles
 	
-Version of org.eclipse.osgi and com.ibm.cds bundles in the above command may not be correct. 
+Version of org.eclipse.osgi and org.openj9.test.cds bundles in the above command may not be correct. 
 Please use the same version as present in cdsadaptortest.jar.
 
-Tests 1-a to 1-e in cdsadaptortest.xml use com.ibm.cdsadaptortest.CDSAdaptorOrphanTest. 
+Tests 1-a to 1-e in cdsadaptortest.xml use org.openj9.test.cdsadaptortest.CDSAdaptorOrphanTest. 
 Please check these tests in the job output for more information.
 
 Eclipse version:
-Eclipse version used to create the two bundles com.ibm.testbundle and com.ibm.weavinghooktest is:
+Eclipse version used to create the two bundles org.openj9.test.testbundle and org.openj9.test.weavinghooktest is:
 Eclipse Juno M5
 Version: 4.2.0
 Build id: I20120127-1145
