+++ test/functional/cmdLineTests/CDSAdaptorTest/src/org/openj9/test/cdsadaptortest/CDSAdaptorOrphanTest.java
@@ -1,13 +1,26 @@
 /*******************************************************************************
- * Licensed Materials - Property of IBM
- * "Restricted Materials of IBM"
+ * Copyright (c) 2001, 2020 IBM Corp. and others
  *
- * (c) Copyright IBM Corp. 2001, 2012 All Rights Reserved
+ * This program and the accompanying materials are made available under
+ * the terms of the Eclipse Public License 2.0 which accompanies this
+ * distribution and is available at https://www.eclipse.org/legal/epl-2.0/
+ * or the Apache License, Version 2.0 which accompanies this distribution and
+ * is available at https://www.apache.org/licenses/LICENSE-2.0.
  *
- * US Government Users Restricted Rights - Use, duplication or disclosure
- * restricted by GSA ADP Schedule Contract with IBM Corp.
+ * This Source Code may also be made available under the following
+ * Secondary Licenses when the conditions for such availability set
+ * forth in the Eclipse Public License, v. 2.0 are satisfied: GNU
+ * General Public License, version 2 with the GNU Classpath
+ * Exception [1] and GNU General Public License, version 2 with the
+ * OpenJDK Assembly Exception [2].
+ *
+ * [1] https://www.gnu.org/software/classpath/license.html
+ * [2] http://openjdk.java.net/legal/assembly-exception.html
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0 OR GPL-2.0 WITH Classpath-exception-2.0 OR LicenseRef-GPL-2.0 WITH Assembly-exception
  *******************************************************************************/
-package com.ibm.cdsadaptortest;
+
+package org.openj9.test.cdsadaptortest;
 
 import org.osgi.framework.*;
 import org.osgi.framework.launch.*;
@@ -23,9 +36,6 @@ import java.util.List;
  * 		-testBundleLocation <location>        - location to install bundles required by the test 
  * Optional arguments: 
  * 		-ignoreWeavingHookBundle              - if present, do not install bundle that performs weaving
- * 
- * @author administrator
- * 
  */
 class CDSAdaptorOrphanTest {
 
@@ -81,7 +91,7 @@ class CDSAdaptorOrphanTest {
 		List<String> excludeList = null;
 		if (ignoreWeavingHookBundle) {
 			excludeList = new ArrayList<String>();
-			excludeList.add("com.ibm.weavinghooktest");
+			excludeList.add("org.openj9.test.weavinghooktest");
 		}
 		List<Bundle> bundleList = launcher.installBundles(testBundleLocation,
 				excludeList);
